✅ STEP1: Avoid type conversions when possible, and make them explicit with a static_cast if you need to do one

+++ 01-The trivia game/Game.cpp
@@ -52,7 +52,7 @@ bool Game::add(const std::string& playerName)
     return true;
 }
 
-int Game::howManyPlayers()
+size_t Game::howManyPlayers()
 {
     return players.size();
 }
